+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 **0.13.0**
 1. Enabling preimages configuration
+2. [eth RPC endpoint]: added zen_dump rpc method for requesting an EVM state dump on a json file
 
 **0.12.0**
 1. Sparkz dependency updated to 2.4.0